fix(llamaindex): refactor event emition logic
